requirements.txt
fastapi
uvicorn
python-dotenv
openai
pillow
pytest
pytesseract
python-multipart